requirements.txt
pymupdf>=1.24.0
python-docx>=1.1.0
Pillow>=10.0.0
